Se finaliza TP1

+++ TP1/Sources/funcionesGet.c
@@ -1,14 +1,6 @@
-/*
- * get_numbers.c
- *
- *  Created on: 13 abr. 2020
- *      Author: julian
- */
-
-
 #include <stdio.h>
 
-long long int getIntValidacion(char mensajeIngreso[], char errorIngreso[], int maximo, int minimo)
+int getIntValidacion(char mensajeIngreso[], char errorIngreso[], int maximo, int minimo)
 {
 	long long int numero;
